refactor: move shouldHandleShortcut into keyboardShortcuts
Its only consumer is the keydown handler, so it belongs beside it rather
than in the shared utilities module. Its tests move with it.
